Add NTAG i2c and bug fix
also streamlined select tag code

+++ client/cmdhf14a.c
@@ -178,7 +178,7 @@ int CmdHF14AReader(const char *Cmd)
 			c.arg[2] = 0;
 			SendCommand(&c);
 			
-			uint16_t tagT = GetHF14AMfU_Type();
+			uint32_t tagT = GetHF14AMfU_Type();
 			ul_print_type(tagT, 0);
 
 			//reconnect for further tests
